# Conflicts:
#	src/main/java/com/otn/controller/LinkTypeController.java
#	src/main/java/com/otn/service/LinkTypeService.java
#	src/main/java/com/otn/service/impl/LinkTypeServiceImpl.java
#	src/main/java/com/otn/webservice/WebServiceFactory.java

+++ src/main/java/com/otn/controller/LinkTypeController.java
@@ -88,9 +88,8 @@ public class LinkTypeController {
     @RequestMapping(value = "/{versionId}/{linkType}/{length}", method = RequestMethod.GET)
     @ResponseStatus(HttpStatus.OK)
     @VersionCheckException
-    public Loss calculateLoss(@PathVariable Long versionId, @PathVariable("linkType") String linkType, @PathVariable
-            Double length) {
-        return new Loss(linkTypeService.calculateLoss(versionId, linkType,length));
+    public Loss calculateLoss(@PathVariable Long versionId, @PathVariable("linkType") String linkType, @PathVariable Long length) {
+        return new Loss(linkTypeService.calculateLoss(versionId, linkType,(float)length));
     }
     private class Loss {
         private DecimalFormat df = new DecimalFormat("0.00");
